+++ src/main/java/io/gravitee/el/spel/SpelExpressionParser.java
@@ -30,11 +30,14 @@ import org.springframework.expression.spel.SpelParserConfiguration;
  */
 public class SpelExpressionParser {
 
-    private static final String EXPRESSION_PREFIX = "{";
+    private static final String EXPRESSION_PREFIX = "{#";
     private static final String EXPRESSION_SUFFIX = "}";
-    private static final String EXPRESSION_REGEX = "\\" + EXPRESSION_PREFIX + "(\\s*[^#T\\(\\s])";
+    // This transforms expressions prefixes from user input : {#, {(, or {T
+    // By prefixes that will be well interpreted by our TemplateParserContext : {##, {#(, or {#T
+    // regular '{' characters won't be interpreted as expression prefixes by EL SpelExpressionParser
+    private static final String EXPRESSION_REGEX = "\\{ *([#T(])";
     private static final Pattern EXPRESSION_REGEX_PATTERN = Pattern.compile(EXPRESSION_REGEX);
-    private static final String EXPRESSION_REGEX_SUBSTITUTE = EXPRESSION_PREFIX + "'" + EXPRESSION_PREFIX + "'" + EXPRESSION_SUFFIX + "$1";
+    private static final String EXPRESSION_REGEX_SUBSTITUTE = EXPRESSION_PREFIX + "$1";
     private static final ParserContext PARSER_CONTEXT = new TemplateParserContext(EXPRESSION_PREFIX, EXPRESSION_SUFFIX);
     private static final org.springframework.expression.spel.standard.SpelExpressionParser EXPRESSION_PARSER = new org.springframework.expression.spel.standard.SpelExpressionParser(
         new SpelParserConfiguration(SpelCompilerMode.MIXED, null)